+++ Genesis/src/utopia/genesis/animation/Animation.scala
@@ -1,9 +1,7 @@
 package utopia.genesis.animation
 
-import utopia.genesis.animation.Animation.{MapAnimation, RepeatingAnimation, ReverseAnimation}
+import utopia.genesis.animation.Animation.{MapAnimation, MergeAnimation, RepeatingAnimation, ReverseAnimation}
 import utopia.genesis.animation.transform.{AnimatedTransform, AnimationWithTransform, TimedAnimationWithTranform, TimedTransform}
-import utopia.genesis.shape.path.{ProjectilePath, SPath}
-import utopia.genesis.shape.shape2D.Point
 
 import scala.concurrent.duration.Duration
 
@@ -12,52 +10,43 @@ import scala.concurrent.duration.Duration
   * @author Mikko Hilpinen
   * @since 11.8.2019, v2.1+
   */
-trait Animation[+A]
+trait Animation[+A] extends AnimationLike[A, Animation]
 {
-	// ABSTRACT	---------------------------
-	
-	/**
-	  * Finds a state of this animation
-	  * @param progress Progress over this animation
-	  * @return The state of this animation at the specified point
-	  */
-	def apply(progress: Double): A
-	
-	
-	// COMPUTED	--------------------------
-	
-	/**
-	  * @return A version of this animation that first progresses faster and then slows down as it nears progress 1.0
-	  */
-	def projectileCurved = CurvedAnimation(this, ProjectilePath())
-	
-	/**
-	  * @return A vesion of this animation that progresses fastest at 50% progress and slowest around 0% and
-	  *         100% progress
-	  */
-	def sPathCurved = curved(SPath.default)
-	
-	/**
-	  * @return A version of this animation that progresses fastest at 50% progress and slowest around 0% and
-	  *         100% progress
-	  */
-	def smoothSPathCurved = curved(SPath.smooth)
-	
-	/**
-	  * @return A version of this animation that progresses fastest at 50% progress and slowest around 0% and
-	  *         100% progress
-	  */
-	def verySmoothSPathCurved = curved(SPath.verySmooth)
-	
-	/**
-	  * @return A copy of this animation that has uses reversed progress
-	  */
-	def reversed: Animation[A] = new ReverseAnimation[A](this)
+	// COMPUTED --------------------------
 	
 	/**
 	  * @return An animation that first plays this animation, and then the reverse version of this animation
 	  */
-	def withReverseAppended: Animation[A] = appendWith(reversed)
+	def withReverseAppended = appendWith(reversed)
+	
+	
+	// IMPLEMENTED  ----------------------
+	
+	/**
+	  * @return A copy of this animation that has uses reversed progress
+	  */
+	override def reversed: Animation[A] = new ReverseAnimation[A](this)
+	
+	/**
+	  * Creates a new animation that repeats this one a number of times
+	  * @param times The number of times this animation is repeated
+	  * @return A new animation
+	  */
+	def repeated(times: Int): Animation[A] = new RepeatingAnimation[A](this, times)
+	
+	/**
+	  * @param curvature A curve animation used for transforming progress% values
+	  * @return A curved version of this animation
+	  */
+	def curved(curvature: AnimationLike[Double, Any]): Animation[A] = CurvedAnimation(this, curvature)
+	
+	/**
+	  * Maps this animation
+	  * @param f A mapping function. Please note that this function will be called multiple times.
+	  * @tparam B Type of map result
+	  * @return An animation that always provides the mapped value
+	  */
+	override def map[B](f: A => B): Animation[B] = new MapAnimation(this)(f)
 	
 	
 	// OTHER	--------------------------
@@ -76,7 +65,6 @@ trait Animation[+A]
 	  */
 	def transformedWith[O >: A, R](transform: AnimatedTransform[O, R]) =
 		AnimationWithTransform.wrap[O, R](this, transform)
-	
 	/**
 	  * @param transform Transformation applied (timed)
 	  * @tparam O Type of transformation origin (must be super type of this animation's result)
@@ -88,50 +76,23 @@ trait Animation[+A]
 		TimedAnimationWithTranform.wrapTimedTransform[O, R](this, transform)
 	
 	/**
-	  * Maps this animation
-	  * @param f A mapping function. Please note that this function will be called multiple times.
-	  * @tparam B Type of map result
-	  * @return An animation that always provides the mapped value
+	  * Merges this animation with another animation
+	  * @param other Another animation
+	  * @param f A merging function
+	  * @tparam B Type of the other animation's result
+	  * @tparam R Type of merge result
+	  * @return A merged animation
 	  */
-	def map[B](f: A => B): Animation[B] = new MapAnimation(this)(f)
-	
-	/**
-	  * Creates a new animation that repeats this one a number of times
-	  * @param times The number of times this animation is repeated
-	  * @return A new animation
-	  */
-	def repeated(times: Int): Animation[A] = new RepeatingAnimation[A](this, times)
-	
-	/**
-	  * @param curvature A curve animation used for transforming progress% values
-	  * @return A curved version of this animation
-	  */
-	def curved(curvature: Animation[Double]): Animation[A] = CurvedAnimation(this, curvature)
-	
-	/**
-	  * @param points Progress mapping points where x represents the original progress and y the mapped progress
-	  * @return A new animation that curves its progress to map through the specified control points
-	  */
-	def curvedWith(points: Seq[Point]) = CurvedAnimation(this, points)
-	
-	/**
-	  * @param control1 First progress control point
-	  * @param control2 Second progress control point
-	  * @return A curved version of this animation where the progress curve follows a cubic bezier determined by
-	  *         (0, 0), control1, control2 and (1, 1) where x represents the original progress and y represents mapped
-	  *         (new animation) progress
-	  */
-	def cubicBezierCurved(control1: Double, control2: Double) =
-		CurvedAnimation(this, control1, control2)
+	def mergeWith[B, R](other: AnimationLike[B, Any])(f: (A, B) => R): Animation[R] =
+		new MergeAnimation[A, B, R](this, other)(f)
 	
 	/**
 	  * Appends another animation to this one
 	  * @param another Another animation
-	  * @param switchAt The progress point at which the animations are switched ]0, 1[ (default = 0.5)
 	  * @tparam B Type of the new animation
 	  * @return A new animation that first plays this animation and then the other
 	  */
-	def appendWith[B >: A](another: Animation[B], switchAt: Double = 0.5) =
+	def appendWith[B >: A](another: AnimationLike[B, Any], switchAt: Double = 0.5) =
 		CombinedAnimation(this, another, switchAt)
 }
 
@@ -151,7 +112,7 @@ object Animation
 	  * @tparam A Type of animation result
 	  * @return An animation that always returns the same result
 	  */
-	def fixed[A](state: A): Animation[A] = new FixedAnimation[A](state)
+	def fixed[A](state: A) = new FixedAnimation[A](state)
 	
 	
 	// NESTED	---------------------------
@@ -159,6 +120,13 @@ object Animation
 	private class MapAnimation[A, +B](original: Animation[A])(f: A => B) extends Animation[B]
 	{
 		override def apply(progress: Double) = f(original(progress))
+	}
+	
+	private class MergeAnimation[O1, O2, R](original1: AnimationLike[O1, Any], original2: AnimationLike[O2, Any])
+	                                       (merge: (O1, O2) => R)
+		extends Animation[R]
+	{
+		override def apply(progress: Double) = merge(original1(progress), original2(progress))
 	}
 	
 	private class RepeatingAnimation[+A](original: Animation[A], repeats: Int) extends Animation[A]
@@ -176,18 +144,5 @@ object Animation
 		override def apply(progress: Double) = original(1 - progress)
 		
 		override def reversed = original
-	}
-	
-	private class FixedAnimation[+A](state: A) extends Animation[A]
-	{
-		override def apply(progress: Double) = state
-		
-		override def reversed = this
-		
-		override def withReverseAppended = this
-		
-		override def repeated(times: Int) = this
-		
-		override def curved(curvature: Animation[Double]) = this
 	}
 }